+++ CHANGES.md
@@ -9,6 +9,7 @@ v1.0.0 (TBD)
   (Issue #21)
 - Fixed `all-shared` target (Issue #22)
 - Fixed memory leaks (Issue #23)
+- Updated `pdfioContentSetDashPattern` to accept `double` values (Issue #25)
 - Fixed some issues identified by a Coverity scan.